fix: change cmd in .air.toml to build instead of run
Prior to this commit, running `make watch` will result in the program running
in the background without any way of killing it other than looking for its PID
and killing that. This changes the `cmd` in the `.air.toml` file to run
`make build` instead of `make run`, as well as changing `make build`'s output
file to `tmp/main` instead of `main`.